Extract .plaintext package
to better manage the growing number of helper classes
