+++ console-runner/src/main/java/info/javaspec/console/plaintext/ConsoleReporter.java
@@ -1,7 +1,8 @@
-package info.javaspec.console;
+package info.javaspec.console.plaintext;
 
 import info.javaspec.Spec;
 import info.javaspec.SpecCollection;
+import info.javaspec.console.Reporter;
 
 import java.io.PrintStream;
 import java.util.ArrayDeque;
@@ -10,7 +11,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-final class ConsoleReporter implements Reporter {
+public final class ConsoleReporter implements Reporter {
   private final PrintStream output;
   private final Deque<ReporterScope> scopes;
   private final EventCounter count;
@@ -85,22 +86,24 @@ final class ConsoleReporter implements Reporter {
 
   @Override
   public void specFailed(Spec spec, AssertionError failure) {
-    this.count.specFailed();
-    scopeForCurrentEvents().specFailed(this.count.numSpecsFailed);
-    this.failures.put(this.count.numSpecsFailed, failure);
+    specFailed(failure);
   }
 
   @Override
   public void specFailed(Spec spec, Exception failure) {
-    this.count.specFailed();
-    scopeForCurrentEvents().specFailed(this.count.numSpecsFailed);
-    this.failures.put(this.count.numSpecsFailed, failure);
+    specFailed(failure);
   }
 
   @Override
   public void specPassed(Spec spec) {
     scopeForCurrentEvents().specPassed();
     this.count.specPassed();
+  }
+
+  private void specFailed(Throwable error) {
+    int howManyHaveFailedNow = this.count.specFailed();
+    scopeForCurrentEvents().specFailed(howManyHaveFailedNow);
+    this.failures.put(howManyHaveFailedNow, error);
   }
 
   private void detailSpecFailure(int referenceNumber, Throwable failure) {
@@ -110,66 +113,5 @@ final class ConsoleReporter implements Reporter {
 
   private ReporterScope scopeForCurrentEvents() {
     return this.scopes.peekLast();
-  }
-
-  private static final class EventCounter {
-    private int numCollectionsStarted;
-    private int numSpecsFailed;
-    private int numSpecsPassed;
-    private int numSpecsTotal;
-
-    public EventCounter() {
-      this.numSpecsFailed = 0;
-      this.numSpecsPassed = 0;
-      this.numSpecsTotal = 0;
-    }
-
-    public void beginCollection() {
-      this.numCollectionsStarted++;
-    }
-
-    public boolean hasFailingSpecs() {
-      return this.numSpecsFailed > 0;
-    }
-
-    public boolean haveAnyEventsOccurred() {
-      return this.numCollectionsStarted > 0
-        || this.numSpecsTotal > 0;
-    }
-
-    public void printSpecTally(PrintStream output) {
-      output.println(String.format(
-        "[Testing complete] Passed: %d, Failed: %d, Total: %d",
-        this.numSpecsPassed,
-        this.numSpecsFailed,
-        this.numSpecsTotal
-      ));
-    }
-
-    public void reset() {
-      this.numSpecsFailed = 0;
-      this.numSpecsPassed = 0;
-      this.numSpecsTotal = 0;
-    }
-
-    public void specStarting() {
-      this.numSpecsTotal++;
-    }
-
-    public void specFailed() {
-      this.numSpecsFailed++;
-    }
-
-    public void specPassed() {
-      this.numSpecsPassed++;
-    }
-  }
-
-  static class RunAlreadyStarted extends IllegalStateException {
-    public RunAlreadyStarted() {
-      super("Tried to start a new run, before finishing the current one."
-        + "  Please call #runFinished, first."
-      );
-    }
   }
 }
